refactor: extract useDynamicLimit hook
