+++ frontend/src/pages/courses/CoursesListPage.tsx
@@ -25,6 +25,7 @@ import { useSnackbar } from 'notistack';
 import ConfirmationDialog from '../../components/ui/ConfirmationDialog';
 import CourseRow from './components/listpage/CourseRow';
 import CourseFilter from './components/listpage/CourseFilter';
+import useDynamicLimit from '../../hooks/useDynamicLimit';
 
 
 
@@ -40,28 +41,6 @@ interface ApiResponse {
   totalCourses: number;
   totalPages: number;
   currentPage: number;
-}
-
-function useDynamicLimit(rowHeight = 56) {
-  const theme = useTheme();
-  const isMobile = useMediaQuery(theme.breakpoints.down('sm'));
-  const [limit, setLimit] = useState(10);
-
-  useEffect(() => {
-    const calculateLimit = () => {
-      const offset = isMobile ? 220 : 280;
-      const availableHeight = window.innerHeight - offset;
-      const rows = Math.floor(availableHeight / rowHeight);
-      const capped = isMobile ? Math.min(rows, 6) : rows;
-      setLimit(capped > 0 ? capped : 4);
-    };
-
-    calculateLimit();
-    window.addEventListener('resize', calculateLimit);
-    return () => window.removeEventListener('resize', calculateLimit);
-  }, [rowHeight, isMobile]);
-
-  return limit;
 }
 
 export default function CoursesListPage() {
